[Refactor] Move warning/deprecated/obsolete/get_traceback to lang.util and hide
