+++ python/taichi/lang/matrix.py
@@ -12,8 +12,8 @@ from taichi.lang.enums import Layout
 from taichi.lang.exception import TaichiSyntaxError
 from taichi.lang.field import Field, ScalarField, SNodeHostAccess
 from taichi.lang.util import (cook_dtype, in_python_scope, python_scope,
-                              taichi_scope, to_numpy_type, to_pytorch_type)
-from taichi.tools.util import warning
+                              taichi_scope, to_numpy_type, to_pytorch_type,
+                              warning)
 from taichi.types import CompoundType, primitive_types
