+++ python/taichi/tools/__init__.py
@@ -13,12 +13,8 @@ __all__ = [
     'imresize',
     'imshow',
     'imwrite',
-    'deprecated',
-    'warning',
     'dump_dot',
     'dot_to_pdf',
-    'obsolete',
     'get_kernel_stats',
-    'get_traceback',
     'set_gdb_trigger',
 ]
